refactor: reduce cyclomatic complexity by extracting auth data application and workspace action handlers
Extract applyLocalAuthData from useLocalAuthState (AuthProvider.tsx) to reduce local auth hook complexity. Move workspace action handlers from useForecastWorkspaceController to a new forecastWorkspaceActions module. Extract TabbedToolbarSelectionStrip to its own file. These changes reduce code duplication, improve readability, and reduce measured cyclomatic complexity of affected hooks while maintaining identical behavior.

- New file: src/components/ForecastWorkspace/forecastWorkspaceActions.tsx
- New file: src/components/IntegratedToolbar/TabbedToolbarSelectionStrip.tsx
- Updated: AuthProvider.tsx (reuse applyLocalAuthData in signIn/signUp)
- Updated: useForecastWorkspaceController.tsx (import extracted handlers)
- Updated: IntegratedToolbar.tsx (import extracted component)
- Fixed: empty arrow function in catch() to surface errors in debug logs
