+++ src/components/ForecastWorkspace/useForecastWorkspaceController.tsx
@@ -24,21 +24,7 @@ import { getOutlookColor } from '../../utils/outlookUtils';
 
 const OUTLOOK_TYPE_ORDER: OutlookType[] = ['tornado', 'wind', 'hail', 'categorical', 'totalSevere', 'day4-8'];
 
-interface ForecastWorkspaceActionParams {
-  dispatch: ReturnType<typeof useDispatch>;
-  onLoad: (file: File) => void;
-  mapRef: React.RefObject<ForecastMapHandle | null>;
-  addToast: AddToastFn;
-  forecastCycle: ReturnType<typeof selectForecastCycle>;
-  currentDay: DayType;
-  canUndo: boolean;
-  canRedo: boolean;
-  tempDate: string;
-  setIsEditingDate: React.Dispatch<React.SetStateAction<boolean>>;
-  setIsPackageDownloading: React.Dispatch<React.SetStateAction<boolean>>;
-  fileInputRef: React.RefObject<HTMLInputElement | null>;
-  handleCancelReset: () => void;
-}
+import { ForecastWorkspaceActionParams, useForecastWorkspaceActionHandlers } from './forecastWorkspaceActions';
 
 export interface ForecastWorkspaceController {
   onSave: () => void;
@@ -108,126 +94,7 @@ interface UseForecastWorkspaceControllerOptions {
   cloudTools?: React.ReactNode;
 }
 
-/** Returns the selected file from a file input change event, or null when none was chosen. */
-const getSelectedFile = (e: React.ChangeEvent<HTMLInputElement>): File | null => {
-  return e.target.files?.[0] ?? null;
-};
-
-/** Returns the neighboring forecast day in the requested direction, or null at the ends. */
-const getAdjacentDay = (currentDay: DayType, offset: -1 | 1): DayType | null => {
-  const nextDay = currentDay + offset;
-  if (nextDay < 1 || nextDay > 8) return null;
-  return nextDay as DayType;
-};
-
-/** Parses the day button's data attribute into a DayType, or null when invalid. */
-const getClickedDay = (e: React.MouseEvent<HTMLButtonElement>): DayType | null => {
-  const day = Number(e.currentTarget.dataset.day);
-  return Number.isNaN(day) ? null : day as DayType;
-};
-
-/** Dispatches an action creator only when the corresponding history state allows it. */
-const dispatchHistoryAction = (
-  dispatch: ReturnType<typeof useDispatch>,
-  isAvailable: boolean,
-  actionCreator: typeof undoLastEdit | typeof redoLastEdit
-) => {
-  if (isAvailable) {
-    dispatch(actionCreator());
-  }
-};
-
-/** Constructs all event-handler callbacks for workspace actions. */
-const useForecastWorkspaceActionHandlers = ({
-  dispatch,
-  onLoad,
-  mapRef,
-  addToast,
-  forecastCycle,
-  currentDay,
-  canUndo,
-  canRedo,
-  tempDate,
-  setIsEditingDate,
-  setIsPackageDownloading,
-  fileInputRef,
-  handleCancelReset,
-}: ForecastWorkspaceActionParams) => {
-  const handlePackageDownload = useCallback(async () => {
-    setIsPackageDownloading(true);
-    try {
-      const mapView = mapRef.current?.getView() ?? { center: [39.8283, -98.5795] as [number, number], zoom: 4 };
-      await downloadGfcPackage(forecastCycle, mapView);
-      addToast('Package downloaded!', 'success');
-    } catch {
-      addToast('Failed to create package.', 'error');
-    } finally {
-      setIsPackageDownloading(false);
-    }
-  }, [mapRef, forecastCycle, addToast, setIsPackageDownloading]);
-
-  const handleFileSelect = useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = getSelectedFile(e);
-    if (file) {
-      onLoad(file);
-    }
-    e.target.value = '';
-  }, [onLoad]);
-
-  const handleLoadClick = useCallback(() => fileInputRef.current?.click(), [fileInputRef]);
-
-  const handleReset = useCallback(() => {
-    dispatch(resetForecasts());
-    handleCancelReset();
-    addToast('All drawings reset', 'info');
-  }, [dispatch, addToast, handleCancelReset]);
-
-  const handleDayChange = useCallback((day: DayType) => dispatch(setForecastDay(day)), [dispatch]);
-
-  const handlePrevDay = useCallback(() => {
-    const previousDay = getAdjacentDay(currentDay, -1);
-    if (previousDay) {
-      dispatch(setForecastDay(previousDay));
-    }
-  }, [dispatch, currentDay]);
-
-  const handleNextDay = useCallback(() => {
-    const nextDay = getAdjacentDay(currentDay, 1);
-    if (nextDay) {
-      dispatch(setForecastDay(nextDay));
-    }
-  }, [dispatch, currentDay]);
-
-  const handleDateSave = useCallback(() => {
-    dispatch(setCycleDate(tempDate));
-    setIsEditingDate(false);
-  }, [dispatch, tempDate, setIsEditingDate]);
-
-  const handleDayButtonClick = useCallback((e: React.MouseEvent<HTMLButtonElement>) => {
-    const day = getClickedDay(e);
-    if (day) {
-      handleDayChange(day);
-    }
-  }, [handleDayChange]);
-
-  const handleToggleLowProbability = useCallback(() => dispatch(toggleLowProbability()), [dispatch]);
-  const handleUndo = useCallback(() => dispatchHistoryAction(dispatch, canUndo, undoLastEdit), [canUndo, dispatch]);
-  const handleRedo = useCallback(() => dispatchHistoryAction(dispatch, canRedo, redoLastEdit), [canRedo, dispatch]);
-
-  return {
-    onUndo: handleUndo,
-    onRedo: handleRedo,
-    onLoadClick: handleLoadClick,
-    onPackageDownload: () => { handlePackageDownload().catch(() => {}); },
-    onDateSave: handleDateSave,
-    onDayButtonClick: handleDayButtonClick,
-    onPrevDay: handlePrevDay,
-    onNextDay: handleNextDay,
-    onToggleLowProbability: handleToggleLowProbability,
-    onReset: handleReset,
-    onFileSelect: handleFileSelect,
-  };
-};
+/* Action handlers moved to forecastWorkspaceActions.tsx */
 
 /** Shared controller for all Forecast workspace layouts. */
 export const useForecastWorkspaceController = ({
